+++ pip_installer/wizard/pip_install.py
@@ -19,9 +19,10 @@
 #    If not, see <http://www.gnu.org/licenses/>.
 #
 ###############################################################################
-from odoo import fields, models
-from odoo.exceptions import ValidationError
+import shlex
 import subprocess
+from odoo import _, fields, models
+from odoo.exceptions import AccessError, ValidationError
 
 
 class PipInstall(models.TransientModel):
@@ -35,11 +36,14 @@ class PipInstall(models.TransientModel):
 
     def action_done(self):
         """Function for executing the pip commands in terminal"""
+        if not self.env.user.has_group('pip_installer.group_pip_installer_admin'):
+            raise AccessError(_('Only Pip Installer Administrators can execute pip commands.'))
         command = self.name
         try:
-            if command.startswith("pip"):
+            if command and command.startswith("pip"):
+                args = shlex.split(command)
                 process = subprocess.Popen(['yes'], stdout=subprocess.PIPE)
-                result = subprocess.run(command, shell=True, check=True,
+                result = subprocess.run(args, check=True,
                                         stdin=process.stdout,
                                         stdout=subprocess.PIPE, text=True)
                 message = self.env['import.message'].create(
@@ -54,8 +58,10 @@ class PipInstall(models.TransientModel):
                         'target': 'new'
                     }
             else:
-                raise ValidationError('Please enter a proper pip command')
+                raise ValidationError(_('Please enter a proper pip command'))
+        except ValueError:
+            raise ValidationError(_('Please enter a proper pip command'))
         except subprocess.CalledProcessError:
-            raise ValidationError('Please enter a proper pip command')
+            raise ValidationError(_('Please enter a proper pip command'))
         except TypeError:
-            raise ValidationError('Please enter a proper pip command')
+            raise ValidationError(_('Please enter a proper pip command'))